requirements.txt
digitalai-release-sdk == 23.1.0rc1
requests >= 2.28.1
boto3 == 1.26.117